requirements.txt
psycopg2-binary
SQLAlchemy
python-dotenv
requests
pandas
exchange_calendars
